feat(types): typed public api
