+++ lib/create-contentful-api.ts
@@ -5,271 +5,46 @@
  * @see Entities
  */
 
-/**
- * The different kinds of top level entities you can find in Contentful
- * @namespace Entities
- */
-
-/**
- * System metadata. See <a href="https://www.contentful.com/developers/docs/references/content-delivery-api/#/introduction/common-resource-attributes">Common Resource Attributes</a> for more details.
- * @memberof Entities
- * @typedef Sys
- * @prop {string} type
- * @prop {string} id
- * @prop {Entities.Link} space
- * @prop {string} createdAt
- * @prop {string} updatedAt
- * @prop {number} revision
- */
-
-/**
- * Link to another entity. See <a href="https://www.contentful.com/developers/docs/concepts/links/">Links</a> for more details.
- * @memberof Entities
- * @typedef Link
- * @prop {string} type - type of this entity. Always link.
- * @prop {string} id
- * @prop {string} linkType - type of this link. If defined, either Entry or Asset
- */
-
-/**
- * @memberof ContentfulClientAPI
- * @typedef {Object} ClientAPI
- * @prop {function} getSpace
- * @prop {function} getContentType
- * @prop {function} getContentTypes
- * @prop {function} getEntry
- * @prop {function} getEntries
- * @prop {function} getAsset
- * @prop {function} getAssets
- * @prop {function} parseEntries
- * @prop {function} sync
- */
-
 import { createRequestConfig } from 'contentful-sdk-core'
+import { AxiosInstance } from 'contentful-sdk-core/dist/types/types'
+import {
+  Asset,
+  AssetCollection,
+  ContentType,
+  ContentTypeCollection,
+  Entry,
+  EntryCollection, LocaleCollection,
+  Space, SyncCollection
+} from './common-types'
+import { GetGlobalOptions } from './create-global-options'
 import pagedSync from './paged-sync'
 import normalizeSelect from './utils/normalize-select'
-import {resolveCircular} from "./utils/resolve-circular";
-
-export interface Asset {
-    sys: Sys;
-    fields: {
-        title: string;
-        description: string;
-        file: {
-            url: string;
-            details: {
-                size: number;
-                image?: {
-                    width: number;
-                    height: number;
-                };
-            };
-            fileName: string;
-            contentType: string;
-        };
-    };
-}
-
-export interface ContentfulCollection<T> {
-    total: number;
-    skip: number;
-    limit: number;
-    items: Array<T>;
-}
-
-export type AssetCollection = ContentfulCollection<Asset>
-
-export interface Entry<T> {
-    sys: Sys;
-    fields: T;
-    update(): Promise<Entry<T>>;
-}
-
-export interface EntryCollection<T> extends ContentfulCollection<Entry<T>> {
-    errors?: Array<any>;
-    includes?: any;
-    stringifySafe(replacer?: any, space?: any): string;
-}
-
-export interface ContentType {
-    sys: Sys;
-    name: string;
-    description: string;
-    displayField: string;
-    fields: Array<Field>;
-}
-
-export type ContentTypeCollection = ContentfulCollection<ContentType>;
-
-export interface Locale {
-    code: string
-    name: string
-    default: boolean
-    fallbackCode: string | null
-    sys: {
-        id: string
-        type: 'Locale'
-        version: number
-    }
-}
-
-export type LocaleCollection = ContentfulCollection<Locale>;
-
-export interface SyncCollection {
-    entries: Array<Entry<any>>;
-    assets: Array<Asset>;
-    deletedEntries: Array<Entry<any>>;
-    deletedAssets: Array<Asset>;
-    nextSyncToken: string;
-    stringifySafe(replacer?: any, space?: any): string;
-}
-
-export interface Sys {
-    type: string;
-    id: string;
-    createdAt: string;
-    updatedAt: string;
-    locale: string;
-    revision?: number;
-    space?: {
-        sys: SpaceLink;
-    };
-    contentType: {
-        sys: ContentTypeLink;
-    };
-}
-
-export interface SpaceLink {
-    type: 'Link';
-    linkType: 'Space';
-    id: string;
-}
-
-export interface ContentTypeLink {
-    type: 'Link';
-    linkType: 'ContentType';
-    id: string;
-}
-
-export interface Field {
-    disabled: boolean;
-    id: string;
-    linkType?: string;
-    localized: boolean;
-    name: string;
-    omitted: boolean;
-    required: boolean;
-    type: FieldType;
-    validations: FieldValidation[];
-    items?: FieldItem;
-}
-
-export type FieldType = 'Symbol' | 'Text' | 'Integer' | 'Number' | 'Date' | 'Boolean' | 'Location' | 'Link' | 'Array' | 'Object' | 'RichText';
-
-export interface FieldValidation {
-    unique?: boolean;
-    size?: {
-        min?: number;
-        max?: number;
-    };
-    regexp?: {
-        pattern: string;
-    };
-    linkMimetypeGroup?: string[];
-    in?: string[];
-    linkContentType?: string[];
-    message?: string;
-    nodes?: {
-        'entry-hyperlink'?: FieldValidation[];
-        'embedded-entry-block'?: FieldValidation[];
-        'embedded-entry-inline'?: FieldValidation[];
-    };
-    enabledNodeTypes?: string[];
-}
-
-export interface FieldItem {
-    type: 'Link' | 'Symbol';
-    validations: FieldValidation[];
-    linkType?: 'Entry' | 'Asset';
-}
-
-
-/**
- * Types of fields found in an Entry
- */
-export namespace EntryFields {
-    export type Symbol = string;
-    export type Text = string;
-    export type Integer = number;
-    export type Number = number;
-    export type Date = string;
-    export type Boolean = boolean;
-    export interface Location {
-        lat: string;
-        lon: string;
-    }
-    export type Link<T> = Asset | Entry<T>;
-    export type Array<T = any> = Symbol[] | Entry<T>[] | Asset[];
-    export type Object<T = any> = T;
-    export interface RichText {
-        data:{};
-        content: RichTextContent[];
-        nodeType: 'document';
-    }
-}
-
-interface RichTextDataTarget {
-    sys: {
-        id: string;
-        type: "Link";
-        "linkType": 'Entry' | 'Asset';
-    };
-}
-
-interface RichTextData {
-    uri?: string;
-    target?: RichTextDataTarget;
-}
-
-type RichTextNodeType = 'text' | 'heading-1' | 'heading-2' | 'heading-3' | 'heading-4' | 'heading-5'
-    | 'heading-6' | 'paragraph' | 'hyperlink' | 'entry-hyperlink' | 'asset-hyperlink'
-    | 'unordered-list' | 'ordered-list' | 'list-item' | 'blockquote' | 'hr' | 'embedded-entry-block'
-    | 'embedded-entry-inline';
-
-interface RichTextContent {
-    data: RichTextData;
-    content?: RichTextContent[]
-    marks: {type: ('bold' | 'underline' | 'code' | 'italic')}[];
-    value?: string;
-    nodeType: RichTextNodeType;
-}
-
-
-export interface Space {
-    sys: Sys;
-    name: string;
-    locales: Array<string>;
-}
+import { resolveCircular } from './utils/resolve-circular'
 
 export interface ContentfulClientApi {
-    readonly version: string;
-    getAsset(id: string, query?: any): Promise<Asset>;
-    getAssets(query?: any): Promise<AssetCollection>;
-    getContentType(id: string): Promise<ContentType>;
-    getContentTypes(query?: any): Promise<ContentTypeCollection>;
-    getEntries<T>(query?: any): Promise<EntryCollection<T>>;
-    getEntry<T>(id: string, query?: any): Promise<Entry<T>>;
-    getSpace(): Promise<Space>;
-    getLocales(): Promise<LocaleCollection>;
-    parseEntries<T>(raw: any): Promise<EntryCollection<T>>;
-    sync(query: any): Promise<SyncCollection>;
+  version: string;
+
+  getAsset(id: string, query?: any): Promise<Asset>;
+
+  getAssets(query?: any): Promise<AssetCollection>;
+
+  getContentType(id: string): Promise<ContentType>;
+
+  getContentTypes(query?: any): Promise<ContentTypeCollection>;
+
+  getEntries<T>(query?: any): Promise<EntryCollection<T>>;
+
+  getEntry<T>(id: string, query?: any): Promise<Entry<T>>;
+
+  getSpace(): Promise<Space>;
+
+  getLocales(): Promise<LocaleCollection>;
+
+  parseEntries<T>(raw: any): EntryCollection<T>;
+
+  sync(query: any): Promise<SyncCollection>;
 }
 
-interface GetConfig {
-    context: "space" | "environment";
-    path: any;
-    config?: any;
-}
 /**
  * Creates API object with methods to access functionality from Contentful's
  * Delivery API
@@ -280,8 +55,14 @@ interface GetConfig {
  * @prop {Function} getGlobalOptions - Link resolver preconfigured with global setting
  * @return {ClientAPI}
  */
-export default function createContentfulApi ({ http, getGlobalOptions }): ContentfulClientApi {
-  const notFoundError = (id) => {
+
+interface CreateContentfulApiParams {
+  http: AxiosInstance,
+  getGlobalOptions: GetGlobalOptions
+}
+
+export default function createContentfulApi ({ http, getGlobalOptions }:CreateContentfulApiParams): ContentfulClientApi {
+  const notFoundError = (id?: string) => {
     const error = new Error('The resource could not be found.')
     /*
       error.sys = {
@@ -311,16 +92,27 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
     throw error
   }
 
-  async function get<T> ({ context, path, config }: GetConfig) : Promise<T> {
-    if (context === 'space') {
-      switchToSpace(http)
-    } else if (context === 'environment') {
-      switchToEnvironment(http)
-    } else {
-      throw new Error('unknown context ')
+  interface GetConfig {
+    context: "space" | "environment";
+    path: string;
+    config?: any;
+  }
+
+  async function get<T> ({ context, path, config }: GetConfig): Promise<T> {
+    let baseUrl = context === 'space'
+      ? getGlobalOptions().spaceBaseUrl
+      : getGlobalOptions().environmentBaseUrl
+
+    if (!baseUrl) {
+      throw new Error('You better define some defaults here')
     }
+
+    if (!baseUrl.endsWith('/')) {
+      baseUrl += '/'
+    }
+
     try {
-      const response = await http.get(path, config)
+      const response = await http.get(baseUrl + path, config)
       return response.data
     } catch (error) {
       errorHandler(error)
@@ -330,7 +122,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
   /**
    * Gets the Space which the client is currently configured to use
    * @memberof ContentfulClientAPI
-   * @return {Promise<Entities.Space>} Promise for a Space
+   * @return {Promise<Space>} Promise for a Space
    * @example
    * const contentful = require('contentful')
    *
@@ -342,7 +134,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const space = await client.getSpace()
    * console.log(space)
    */
-  async function getSpace ():Promise<Space> {
+  async function getSpace (): Promise<Space> {
     return get<Space>({ context: 'space', path: '' })
   }
 
@@ -350,7 +142,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * Gets a Content Type
    * @memberof ContentfulClientAPI
    * @param  {string} id
-   * @return {Promise<Entities.ContentType>} Promise for a Content Type
+   * @return {Promise<ContentType>} Promise for a Content Type
    * @example
    * const contentful = require('contentful')
    *
@@ -362,7 +154,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const contentType = await client.getContentType('<content_type_id>')
    * console.log(contentType)
    */
-  async function getContentType (id): Promise<ContentType> {
+  async function getContentType (id: string): Promise<ContentType> {
     return get<ContentType>({
       context: 'environment',
       path: `content_types/${id}`
@@ -373,7 +165,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * Gets a collection of Content Types
    * @memberof ContentfulClientAPI
    * @param  {Object=} query - Object with search parameters. Check the <a href="https://www.contentful.com/developers/docs/javascript/tutorials/using-js-cda-sdk/#retrieving-entries-with-search-parameters">JS SDK tutorial</a> and the <a href="https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/search-parameters">REST API reference</a> for more details.
-   * @return {Promise<Entities.ContentTypeCollection>} Promise for a collection of Content Types
+   * @return {Promise<ContentTypeCollection>} Promise for a collection of Content Types
    * @example
    * const contentful = require('contentful')
    *
@@ -385,7 +177,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const response = await client.getContentTypes()
    * console.log(response.items)
    */
-  async function getContentTypes (query = {}):Promise<ContentTypeCollection> {
+  async function getContentTypes (query = {}): Promise<ContentTypeCollection> {
     return get<ContentTypeCollection>({
       context: 'environment',
       path: 'content_types',
@@ -398,7 +190,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * @memberof ContentfulClientAPI
    * @param  {string} id
    * @param  {Object=} query - Object with search parameters. In this method it's only useful for `locale`.
-   * @return {Promise<Entities.Entry>} Promise for an Entry
+   * @return {Promise<Entry>} Promise for an Entry
    * @example
    * const contentful = require('contentful')
    *
@@ -410,7 +202,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const entry = await client.getEntry('<entry_id>')
    * console.log(entry)
    */
-  async function getEntry <T> (id, query = {}):Promise<Entry<T>> {
+  async function getEntry<T> (id: string, query = {}): Promise<Entry<T>> {
     if (!id) {
       throw notFoundError(id)
     }
@@ -442,7 +234,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const response = await client.getEntries()
    * .console.log(response.items)
    */
-  async function getEntries <T> (query = {}):Promise<EntryCollection<T>> {
+  async function getEntries<T> (query = {}): Promise<EntryCollection<T>> {
     const { resolveLinks, removeUnresolved } = getGlobalOptions(query)
     try {
       const entries = await get({
@@ -455,6 +247,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
       errorHandler(error)
     }
   }
+
   /**
    * Gets an Asset
    * @memberof ContentfulClientAPI
@@ -472,7 +265,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const asset = await client.getAsset('<asset_id>')
    * console.log(asset)
    */
-  async function getAsset (id, query = {}): Promise<Asset> {
+  async function getAsset (id: string, query = {}): Promise<Asset> {
     return get<Asset>({
       context: 'environment',
       path: `assets/${id}`,
@@ -520,7 +313,7 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
    * const response = await client.getLocales()
    * console.log(response.items)
    */
-  async function getLocales (query = {}):Promise<LocaleCollection> {
+  async function getLocales (query = {}): Promise<LocaleCollection> {
     return get<LocaleCollection>({
       context: 'environment',
       path: 'locales',
@@ -568,35 +361,35 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
   }
 
   /**
-  * Parse raw json data into collection of entry objects.Links will be resolved also
-  * @memberof ContentfulClientAPI
-  * @param {Object} raw json data
-  * @example
-  * let data = {items: [
-  *    {
-  *    sys: {type: 'Entry', locale: 'en-US'},
-  *    fields: {
-  *      animal: {sys: {type: 'Link', linkType: 'Animal', id: 'oink'}},
-  *      anotheranimal: {sys: {type: 'Link', linkType: 'Animal', id: 'middle-parrot'}}
-  *    }
-  *  }
-  * ],
-  * includes: {
-  *  Animal: [
-  *    {
-  *      sys: {type: 'Animal', id: 'oink', locale: 'en-US'},
-  *      fields: {
-  *        name: 'Pig',
-  *        friend: {sys: {type: 'Link', linkType: 'Animal', id: 'groundhog'}}
-  *      }
-  *    }
-  *   ]
-  *  }
-  * }
-  * console.log( data.items[0].fields.foo ); // undefined
-  * let parsedData = client.parseEntries(data);
-  * console.log( parsedData.items[0].fields.foo ); // foo
-  */
+   * Parse raw json data into collection of entry objects.Links will be resolved also
+   * @memberof ContentfulClientAPI
+   * @param {Object} raw json data
+   * @example
+   * let data = {items: [
+   *    {
+   *    sys: {type: 'Entry', locale: 'en-US'},
+   *    fields: {
+   *      animal: {sys: {type: 'Link', linkType: 'Animal', id: 'oink'}},
+   *      anotheranimal: {sys: {type: 'Link', linkType: 'Animal', id: 'middle-parrot'}}
+   *    }
+   *  }
+   * ],
+   * includes: {
+   *  Animal: [
+   *    {
+   *      sys: {type: 'Animal', id: 'oink', locale: 'en-US'},
+   *      fields: {
+   *        name: 'Pig',
+   *        friend: {sys: {type: 'Link', linkType: 'Animal', id: 'groundhog'}}
+   *      }
+   *    }
+   *   ]
+   *  }
+   * }
+   * console.log( data.items[0].fields.foo ); // undefined
+   * let parsedData = client.parseEntries(data);
+   * console.log( parsedData.items[0].fields.foo ); // foo
+   */
   function parseEntries (data) {
     const { resolveLinks, removeUnresolved } = getGlobalOptions({})
     return resolveCircular(data, { resolveLinks, removeUnresolved })
@@ -605,15 +398,8 @@ export default function createContentfulApi ({ http, getGlobalOptions }): Conten
   /*
    * Switches BaseURL to use /environments path
    * */
-  function switchToEnvironment (http): void {
+  function switchToEnvironment (http: AxiosInstance): void {
     http.defaults.baseURL = getGlobalOptions().environmentBaseUrl
-  }
-
-  /*
-   * Switches BaseURL to use /spaces path
-   * */
-  function switchToSpace (http): void {
-    http.defaults.baseURL = getGlobalOptions().spaceBaseUrl
   }
 
   return {
