+++ package.json
@@ -63,6 +63,8 @@
     "@babel/register": "^7.6.2",
     "@contentful/browserslist-config": "^1.0.1",
     "@semantic-release/changelog": "^5.0.1",
+    "@types/axios": "^0.14.0",
+    "@types/json-stringify-safe": "^5.0.0",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "babel-plugin-inline-replace-variables": "^1.3.1",
